refactor: 클래스 선언부 공백 처리
