+++ src/main/java/dnd/myOcean/config/oAuth/KakaoConfig.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Component;
 @Component
 @Getter
 public class KakaoConfig {
+
     @Value("${spring.security.oauth2.client.registration.kakao.client-id}")
     private String kakaoClientId;
