+++ src/main/java/dnd/myOcean/domain/member/Role.java
@@ -1,4 +1,4 @@
-package dnd.myOcean.util;
+package dnd.myOcean.domain.member;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 @Getter
 @RequiredArgsConstructor
 public enum Role {
+
     USER("유저"), ADMIN("관리자");
 
     private final String authority;
