+++ src/main/java/dnd/myOcean/dto/jwt/response/TokenDto.java
@@ -8,6 +8,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 public class TokenDto {
+
     private final String accessToken;
     private final String refreshToken;
 }
