+++ src/main/java/dnd/myOcean/dto/oAuth/response/MemberInfo.java
@@ -8,6 +8,7 @@ import lombok.Getter;
 @Builder
 @AllArgsConstructor
 public class MemberInfo {
+
     private final Long id;
     private final String email;
     private final String nickName;
